# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Library App using Golang",
        "title": "Library App",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/v1/book": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book API"
                ],
                "summary": "Get a list of books",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apis.PaginationResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/apis.BookResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "client error",
                        "schema": {
                            "$ref": "#/definitions/apis.ExceptionResponse"
                        }
                    },
                    "500": {
                        "description": "database error",
                        "schema": {
                            "$ref": "#/definitions/apis.ExceptionResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book API"
                ],
                "summary": "Create one book",
                "parameters": [
                    {
                        "description": "Create book request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.CreateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apis.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apis.BookResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "client error",
                        "schema": {
                            "$ref": "#/definitions/apis.ExceptionResponse"
                        }
                    },
                    "500": {
                        "description": "database error",
                        "schema": {
                            "$ref": "#/definitions/apis.ExceptionResponse"
                        }
                    }
                }
            }
        },
        "/v1/book/borrow": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book API"
                ],
                "summary": "Borrow a book",
                "parameters": [
                    {
                        "description": "Borrow book request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.BorrowBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apis.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apis.BorrowBookResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "client error",
                        "schema": {
                            "$ref": "#/definitions/apis.ExceptionResponse"
                        }
                    },
                    "500": {
                        "description": "database error",
                        "schema": {
                            "$ref": "#/definitions/apis.ExceptionResponse"
                        }
                    }
                }
            }
        },
        "/v1/book/borrow/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book API"
                ],
                "summary": "Get list of books by borrowId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "borrower id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apis.PaginationResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apis.BookResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "client error",
                        "schema": {
                            "$ref": "#/definitions/apis.ExceptionResponse"
                        }
                    },
                    "500": {
                        "description": "database error",
                        "schema": {
                            "$ref": "#/definitions/apis.ExceptionResponse"
                        }
                    }
                }
            }
        },
        "/v1/book/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book API"
                ],
                "summary": "Get one book by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "book id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apis.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apis.BookResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "client error",
                        "schema": {
                            "$ref": "#/definitions/apis.ExceptionResponse"
                        }
                    },
                    "500": {
                        "description": "database error",
                        "schema": {
                            "$ref": "#/definitions/apis.ExceptionResponse"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book API"
                ],
                "summary": "Delete one book by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "book id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apis.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apis.BookResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "client error",
                        "schema": {
                            "$ref": "#/definitions/apis.ExceptionResponse"
                        }
                    },
                    "500": {
                        "description": "database error",
                        "schema": {
                            "$ref": "#/definitions/apis.ExceptionResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book API"
                ],
                "summary": "Update one book by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "book id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update book request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.UpdateOneBookByIdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apis.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apis.BookResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "client error",
                        "schema": {
                            "$ref": "#/definitions/apis.ExceptionResponse"
                        }
                    },
                    "500": {
                        "description": "database error",
                        "schema": {
                            "$ref": "#/definitions/apis.ExceptionResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apis.BookResponse": {
            "type": "object",
            "properties": {
                "book": {
                    "$ref": "#/definitions/db.Book"
                },
                "book_info": {
                    "$ref": "#/definitions/db.BookInfo"
                },
                "category": {
                    "$ref": "#/definitions/db.Category"
                }
            }
        },
        "apis.BorrowBookRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "book_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "apis.BorrowBookResponse": {
            "type": "object",
            "properties": {
                "book": {
                    "$ref": "#/definitions/db.Book"
                },
                "book_info": {
                    "$ref": "#/definitions/db.BookInfo"
                },
                "borrower": {
                    "$ref": "#/definitions/db.Borrower"
                },
                "category": {
                    "$ref": "#/definitions/db.Category"
                }
            }
        },
        "apis.CreateBookRequest": {
            "type": "object",
            "required": [
                "author",
                "category_id",
                "name",
                "publication_date"
            ],
            "properties": {
                "author": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "publication_date": {
                    "type": "string"
                }
            }
        },
...
        "apis.ExceptionResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/validations.ValidationError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "apis.PaginatedResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "apis.PaginationResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/apis.PaginatedResponse"
                },
                "status_code": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "apis.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "apis.UpdateOneBookByIdRequest": {
            "type": "object",
            "required": [
                "author",
                "category_id",
                "name",
                "publication_date"
            ],
            "properties": {
                "author": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "publication_date": {
                    "type": "string"
                }
            }
        },
...
        "db.Book": {
            "type": "object",
            "properties": {
                "book_info_id": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "db.BookInfo": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "publication_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "db.Borrower": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "db.Category": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "validations.ValidationError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                }
            }
        }
    }
}